[#69] Add Notifier trait with example
