+++ iceoryx2/src/prelude.rs
@@ -12,5 +12,5 @@
 
 pub use crate::iox2::Iox2;
 pub use crate::iox2::Iox2Event;
-pub use crate::port::{event_id::EventId, listener::Listener};
+pub use crate::port::{event_id::EventId, listener::Listener, notifier::Notifier};
 pub use crate::service::{process_local, service_name::ServiceName, zero_copy, Details, Service};
